extract redis oplog wrapper into separate file
