+++ lib/hijack/instrument.js
@@ -12,11 +12,11 @@ import {
   wrapMultiplexer,
   wrapOplogObserveDriver,
   wrapPollingObserveDriver,
-  wrapRedisOplogObserveDriver
 } from './wrap_observers';
 import { wrapStringifyDDP } from './wrap_ddp_stringify';
 import { setLabels } from './set_labels';
 import { hijackDBOps } from './db';
+import { wrapRedisOplogObserveDriver } from './redis_oplog';
 
 let instrumented = false;
 Kadira._startInstrumenting = function (callback) {
